package refactoring, 3 task start ralization

+++ src/main/java/com/elementary/io/DoubleConsoleWorker.java
@@ -2,13 +2,13 @@ package com.elementary.io;
 
 import java.util.Optional;
 
-public class DoubleConsoleWorker extends ComparableConsoleWorker<Float, DoubleConsoleWorker> {
+public class DoubleConsoleWorker extends ComparableConsoleWorker<Double, DoubleConsoleWorker> {
 
     @Override
-    protected Optional<Float> parse(String input) {
-        Float value;
+    protected Optional<Double> parse(String input) {
+        Double value;
         try {
-            value = Float.parseFloat(input);
+            value = Double.parseDouble(input);
         } catch (NumberFormatException e) {
             value = null;
         }
